Fix warnings

+++ app/common/view.c
@@ -332,7 +332,7 @@ zxerr_t h_review_update_data() {
         viewdata.itemCount++;
 
         if (viewdata.pageCount > 1) {
-            uint8_t keyLen = strnlen(viewdata.key, MAX_CHARS_PER_KEY_LINE);
+            uint8_t keyLen = (uint8_t) strnlen(viewdata.key, MAX_CHARS_PER_KEY_LINE);
             if (keyLen < MAX_CHARS_PER_KEY_LINE) {
                 snprintf(viewdata.key + keyLen,
                          MAX_CHARS_PER_KEY_LINE - keyLen,
